requirements.txt
torch
transformers
accelerate
peft
sentencepiece
safetensors
pyyaml
rouge-score
tensorboard
datasets
tqdm
matplotlib
huggingface_hub
streamlit
